Added deleting mechanism for users posts.
To achieve this, we needed to reference the document ids, which I needed
to restructure saving that within the firebase documents. This mechanism
will allow for dynamic slugs for individual posts in the future. The
delete mechanism shows up if the logged in UID matches the UID in the post