Added dynamic routing for detailed posts.
The detailed post card adds parts of the post that are not
in the initial card that users use. Things like square footage,
number of roommates, and a description are included. To achieve this,
we used Next.js dynamic routing to route to the document id for firebase
that contains the post information.

Using this piece of information, we can call the firebase api to get the
document that matches the piece of information.