Hotfix for card image sizing and typo in post detail.
For the card component, the image sizes were having issues and
rendering differently for images with different aspect ratios.
In the new patch, I added a static height to 48pt which will allow for
all of the images to stay the same size.

In the post detail component, I caught a spelling error in the word
"roommates" and fixed it.